feat(js): split classes to controller-model-view folder setup (object-oriented)
