+++ index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
   <canvas id="gameCanvas" width="700" height="700"></canvas>
-  <script src="main.js"></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
